renamed Hooks to hooks, moved userContext to 'context' folder
